add text_bulletinboard.py
